# dashboard/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: MONOLITH OS API
  description: |
    RESTful API for MONOLITH OS - An AI-powered enterprise operating system that orchestrates
    autonomous agents for business operations, decision-making, and workflow management.

    ## Features
    - **Task Management**: Create, update, and track tasks across the organization
    - **Decision Logging**: Record and audit business decisions with full traceability
    - **Workflow Orchestration**: Manage multi-step business processes
    - **Agent Management**: Monitor and control AI agents across different roles

    ## Authentication
    All endpoints require Bearer token authentication. Obtain a token via `/auth/login`.

    ## Rate Limiting
    - Standard endpoints: 100 requests/minute
    - Bulk operations: 10 requests/minute

    ## Versioning
    The API uses URL-based versioning. Current version: v1
  version: 1.0.0
  contact:
    name: MONOLITH OS Support
    email: support@monolith-os.local
  license:
    name: Proprietary
    url: https://monolith-os.local/license
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.monolith-os.local/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login
      description: Authenticate and receive a JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register
      description: Register a new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                  minLength: 8
                name:
                  type: string
                role:
                  type: string
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh token
      description: Refresh an expired access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            field:
              type: string
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            expiresIn:
              type: integer
            user:
              type: object
              properties:
                id:
                  type: string
                email:
                  type: string
                name:
                  type: string
                role:
                  type: string
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid input
              field: content
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
...